+++ test/micro_test.rb
@@ -1,6 +1,16 @@
 require 'test_helper'
 
 class MicroTest < Minitest::Test
+  def test_attr_reader
+    assert_change 'def foo; @foo; end', 'attr_reader :foo'
+    refute_change "def foo \n@bar\nend"
+  end
+
+  def test_attr_writer
+    assert_change 'def foo=(value); @foo = value; end', 'attr_writer :foo'
+    refute_change "def foo= (value)\n@bar = value\nend"
+  end
+
   def test_reverse_each
     assert_change '[].reverse.each', '[].reverse_each'
     assert_change 'foo.reverse.each', 'foo.reverse_each'